Module BantorraWebResolvers.Git

The Builder

val resolver : strict_checking:bool -> crate_root:string -> Bantorra.Resolver.t

resolver ~crate_root downloads git repositories, when requested, into the directory crate_root and resolves the URL into a path pointing to a local copy of the library. The resolver assumes the directory crate_root already exists and it has total control over the directory during the program execution.

parameter strict_checking

Whether full resolution is performed to check the validity of arguments. If the value is false, the resolver will only check whether the arguments are well-formed.